Add 4-way parallel queue
Refactor data passing between plugin and tag so it even _can_ run in parallel. Prev versions used a context store that is shared for the request, so multiple runs for the same request would clobber each other's stores. Now it uses an extraData param on the sendRequest() call, which is isolated